requirements.txt
# Document Processing
pymupdf4llm>=0.0.17
pymupdf>=1.24.0
python-docx>=1.1.0

# Vector Database
chromadb>=0.5.0

# Embeddings & LLM
openai>=1.30.0
anthropic>=0.40.0
groq>=0.9.0
google-genai>=1.0.0

# UI
gradio>=5.0.0

# Citations
pybtex>=0.24.0

# MCP Server
mcp>=1.0.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
tqdm>=4.66.0
